+++ src/CirculationLogList/CirculationLogListFilter/TextFilters/TextFilters.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { act, render } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import { byRole } from 'testing-library-selector';
 
 import { Pluggable } from '@folio/stripes/core';
@@ -13,13 +14,18 @@ jest.mock('@folio/stripes/core', () => ({
 
 describe('"Find user" plugin', () => {
   const TYPE = 'find-user';
+  const fields = {
+    user: byRole('textbox', { name: /user/ }),
+    description: byRole('textbox', { name: /description/ }),
+  };
 
-  it('is invoked with `selectUser` callback in props', () => {
+  it('is invoked with `selectUser` and `afterClose` callbacks in props', () => {
     render(<TextFilters activeFilters={{}} applyFilters={() => {}} />);
 
     const expectedProps = expect.objectContaining({
       type: TYPE,
       selectUser: expect.any(Function),
+      afterClose: expect.any(Function),
     });
     const expectedContext = expect.anything();
 
@@ -27,8 +33,6 @@ describe('"Find user" plugin', () => {
   });
 
   it('can use `selectUser` callback to set the "User barcode" field value', () => {
-    const field = byRole('textbox', { name: /user/ });
-    const user = { barcode: 'test barcode' };
     let callback = () => {};
 
     Pluggable.mockImplementation(({ type, selectUser }) => {
@@ -39,10 +43,34 @@ describe('"Find user" plugin', () => {
 
     render(<TextFilters activeFilters={{}} applyFilters={() => {}} />);
 
-    expect(field.get()).toHaveValue('');
+    const userField = fields.user.get();
 
-    act(() => callback(user));
+    expect(userField).toHaveValue('');
 
-    expect(field.get()).toHaveValue('test barcode');
+    act(() => callback({ barcode: 'test barcode' }));
+
+    expect(userField).toHaveValue('test barcode');
+  });
+
+  it('can use `afterClose` callback to set focus to the "User barcode" field', () => {
+    let callback = () => {};
+
+    Pluggable.mockImplementation(({ type, afterClose }) => {
+      if (type === TYPE) callback = afterClose;
+
+      return 'plugin';
+    });
+
+    render(<TextFilters activeFilters={{}} applyFilters={() => {}} />);
+
+    const userField = fields.user.get();
+
+    // Focusing out of userField
+    userEvent.click(fields.description.get());
+    expect(userField).not.toHaveFocus();
+
+    act(callback);
+
+    expect(userField).toHaveFocus();
   });
 });